@@ -25,6 +25,7 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
+  aligned_dialog: 0.0.6
   auto_size_text: 3.0.0
   cached_network_image: 3.2.1
   csv: 5.0.1
@@ -32,31 +33,33 @@ dependencies:
   equatable: 2.0.5
   fl_chart: 0.55.0
   flutter_animate: 4.1.1+1
-  flutter_cache_manager: 3.3.0
+  flutter_cache_manager: 3.3.1
   flutter_native_splash: 2.3.1
   flutter_secure_storage: 8.0.0
-  flutter_spinkit: 5.1.0
-  font_awesome_flutter: 10.1.0
+  flutter_spinkit: 5.2.0
+  flutter_staggered_grid_view: 0.7.0
+  font_awesome_flutter: 10.6.0
   from_css_color: 2.0.0
   go_router: 7.1.1
   google_fonts: 4.0.3
   intl: 0.18.1
   json_path: 0.6.2
   mime_type: 1.0.0
-  page_transition: 2.0.4
+  page_transition: 2.1.0
   path_provider: 2.0.14
   path_provider_android: 2.0.25
   path_provider_foundation: 2.2.2
   path_provider_platform_interface: 2.0.6
   percent_indicator: 4.2.2
-  provider: 6.0.4
+  photo_view: 0.14.0
+  provider: 6.0.5
   purchases_flutter: 6.0.0
-  shared_preferences: 2.0.15
-  shared_preferences_android: 2.1.0
-  shared_preferences_ios: 2.1.1
-  shared_preferences_platform_interface: 2.2.0
-  shared_preferences_web: 2.1.0
-  smooth_page_indicator: 1.0.1
+  shared_preferences: 2.2.2
+  shared_preferences_android: 2.2.1
+  shared_preferences_foundation: 2.3.4
+  shared_preferences_platform_interface: 2.3.1
+  shared_preferences_web: 2.2.1
+  smooth_page_indicator: 1.1.0
   sqflite: 2.2.6
   synchronized: 3.1.0
   timeago: 3.2.2
