@@ -1,5 +1,5 @@
+import '/components/achievement_category_widget.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/pages/dashboard/dashboard_tab_page/dashboard_tab_page_widget.dart';
 import '/pages/generic/app_drawer/app_drawer_widget.dart';
 import '/pages/generic/base_header/base_header_widget.dart';
 import 'dashboard_widget.dart' show DashboardWidget;
@@ -9,21 +9,8 @@ class DashboardModel extends FlutterFlowModel<DashboardWidget> {
   ///  State fields for stateful widgets in this page.
 
   final unfocusNode = FocusNode();
-  // State field(s) for TabBar widget.
-  TabController? tabBarController;
-  int get tabBarCurrentIndex =>
-      tabBarController != null ? tabBarController!.index : 0;
-
-  // Model for dashboardTabPage component.
-  late DashboardTabPageModel dashboardTabPageModel1;
-  // Model for dashboardTabPage component.
-  late DashboardTabPageModel dashboardTabPageModel2;
-  // Model for dashboardTabPage component.
-  late DashboardTabPageModel dashboardTabPageModel3;
-  // Model for dashboardTabPage component.
-  late DashboardTabPageModel dashboardTabPageModel4;
-  // Model for dashboardTabPage component.
-  late DashboardTabPageModel dashboardTabPageModel5;
+  // Model for AchievementCategory component.
+  late AchievementCategoryModel achievementCategoryModel;
   // Model for baseHeader component.
   late BaseHeaderModel baseHeaderModel;
   // Model for appDrawer component.
@@ -33,16 +20,8 @@ class DashboardModel extends FlutterFlowModel<DashboardWidget> {
 
   @override
   void initState(BuildContext context) {
-    dashboardTabPageModel1 =
-        createModel(context, () => DashboardTabPageModel());
-    dashboardTabPageModel2 =
-        createModel(context, () => DashboardTabPageModel());
-    dashboardTabPageModel3 =
-        createModel(context, () => DashboardTabPageModel());
-    dashboardTabPageModel4 =
-        createModel(context, () => DashboardTabPageModel());
-    dashboardTabPageModel5 =
-        createModel(context, () => DashboardTabPageModel());
+    achievementCategoryModel =
+        createModel(context, () => AchievementCategoryModel());
     baseHeaderModel = createModel(context, () => BaseHeaderModel());
     appDrawerModel = createModel(context, () => AppDrawerModel());
   }
@@ -50,12 +29,7 @@ class DashboardModel extends FlutterFlowModel<DashboardWidget> {
   @override
   void dispose() {
     unfocusNode.dispose();
-    tabBarController?.dispose();
-    dashboardTabPageModel1.dispose();
-    dashboardTabPageModel2.dispose();
-    dashboardTabPageModel3.dispose();
-    dashboardTabPageModel4.dispose();
-    dashboardTabPageModel5.dispose();
+    achievementCategoryModel.dispose();
     baseHeaderModel.dispose();
     appDrawerModel.dispose();
   }
