@@ -25,6 +25,7 @@ class ApiGroup {
   static RankingsCall rankingsCall = RankingsCall();
   static AdvicesCall advicesCall = AdvicesCall();
   static ContestsCall contestsCall = ContestsCall();
+  static GetContestsVTwoCall getContestsVTwoCall = GetContestsVTwoCall();
   static DashboardGraphCall dashboardGraphCall = DashboardGraphCall();
   static UpdateUserPropertyCall updateUserPropertyCall =
       UpdateUserPropertyCall();
@@ -39,6 +40,11 @@ class ApiGroup {
   static MarkRecomendationAsReadCall markRecomendationAsReadCall =
       MarkRecomendationAsReadCall();
   static SwimTimeToClubIdCall swimTimeToClubIdCall = SwimTimeToClubIdCall();
+  static GetAchievementCategoriesCall getAchievementCategoriesCall =
+      GetAchievementCategoriesCall();
+  static GetAchievementsInCategoryCall getAchievementsInCategoryCall =
+      GetAchievementsInCategoryCall();
+  static PremiumStateCall premiumStateCall = PremiumStateCall();
 }
 
 class DashboardCall {
@@ -61,38 +67,39 @@ class DashboardCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
-  dynamic isValid(dynamic response) => getJsonField(
+  bool? isValid(dynamic response) => castToType<bool>(getJsonField(
         response,
         r'''$.data.linkedAccount''',
-      );
-  dynamic freestylePerformances(dynamic response) => getJsonField(
+      ));
+  List? freestylePerformances(dynamic response) => getJsonField(
         response,
         r'''$.data.performances.freeStyle''',
         true,
-      );
-  dynamic butterflyPerformances(dynamic response) => getJsonField(
+      ) as List?;
+  List? butterflyPerformances(dynamic response) => getJsonField(
         response,
         r'''$.data.performances.butterfly''',
         true,
-      );
-  dynamic backstrokePerformances(dynamic response) => getJsonField(
+      ) as List?;
+  List? backstrokePerformances(dynamic response) => getJsonField(
         response,
         r'''$.data.performances.backstroke''',
         true,
-      );
-  dynamic breaststrokePerformances(dynamic response) => getJsonField(
+      ) as List?;
+  List? breaststrokePerformances(dynamic response) => getJsonField(
         response,
         r'''$.data.performances.breaststroke''',
         true,
-      );
-  dynamic medleyPerformances(dynamic response) => getJsonField(
+      ) as List?;
+  List? medleyPerformances(dynamic response) => getJsonField(
         response,
         r'''$.data.performances.medley''',
         true,
-      );
+      ) as List?;
 }
 
 class CreateUserByDeviceIdentifierCall {
@@ -118,22 +125,23 @@ class CreateUserByDeviceIdentifierCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
-  dynamic exists(dynamic response) => getJsonField(
+  bool? exists(dynamic response) => castToType<bool>(getJsonField(
         response,
         r'''$.data.exists''',
-      );
+      ));
   dynamic activeUser(dynamic response) => getJsonField(
         response,
         r'''$.data.active''',
       );
-  dynamic favorites(dynamic response) => getJsonField(
+  List? favorites(dynamic response) => getJsonField(
         response,
         r'''$.data.favorites''',
         true,
-      );
+      ) as List?;
 }
 
 class AddFavoritedUserCall {
@@ -159,6 +167,7 @@ class AddFavoritedUserCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 }
@@ -184,6 +193,7 @@ class SetActiveUserCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 }
@@ -193,6 +203,8 @@ class RaceListCall {
     String? deviceIdentifier = '',
     String? style = '',
     String? course = '',
+    String? raceId = '-1',
+    String? year = '-1',
   }) async {
     return ApiManager.instance.makeApiCall(
       callName: 'raceList',
@@ -205,32 +217,64 @@ class RaceListCall {
         'deviceIdentifier': deviceIdentifier,
         'style': style,
         'course': course,
+        'raceId': raceId,
+        'year': year,
       },
       returnBody: true,
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
-  dynamic races(dynamic response) => getJsonField(
+  List? races(dynamic response) => getJsonField(
         response,
         r'''$.data.races''',
         true,
-      );
-  dynamic rankings(dynamic response) => getJsonField(
+      ) as List?;
+  List? rankings(dynamic response) => getJsonField(
         response,
         r'''$.data.rankings''',
         true,
-      );
-  dynamic max(dynamic response) => getJsonField(
+      ) as List?;
+  int? max(dynamic response) => castToType<int>(getJsonField(
         response,
         r'''$.data.maxScore''',
-      );
-  dynamic scoreTime(dynamic response) => getJsonField(
+      ));
+  String? scoreTime(dynamic response) => castToType<String>(getJsonField(
         response,
         r'''$.data.maxScoreTime''',
+      ));
+  bool? hasRankings(dynamic response) => castToType<bool>(getJsonField(
+        response,
+        r'''$.data.hasRankings''',
+      ));
+  dynamic achievement(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement''',
       );
+  dynamic achievementLevel(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.level.level''',
+      );
+  dynamic achievementLevelName(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.level.name''',
+      );
+  dynamic achievementLebelBadge(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.level.badge''',
+      );
+  dynamic achievementLevelCount(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.levelCount''',
+      );
+  List? achievementLevels(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.levels''',
+        true,
+      ) as List?;
 }
 
 class DeleteFavoritedUserCall {
@@ -254,6 +298,7 @@ class DeleteFavoritedUserCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 }
@@ -280,6 +325,7 @@ class RankingsCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
@@ -291,11 +337,11 @@ class RankingsCall {
         response,
         r'''$.data.toResult''',
       );
-  dynamic results(dynamic response) => getJsonField(
+  List? results(dynamic response) => getJsonField(
         response,
         r'''$.data.results''',
         true,
-      );
+      ) as List?;
 }
 
 class AdvicesCall {
@@ -318,14 +364,15 @@ class AdvicesCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
-  dynamic advices(dynamic response) => getJsonField(
+  List? advices(dynamic response) => getJsonField(
         response,
         r'''$.data.advices''',
         true,
-      );
+      ) as List?;
 }
 
 class ContestsCall {
@@ -346,8 +393,66 @@ class ContestsCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
+}
+
+class GetContestsVTwoCall {
+  Future<ApiCallResponse> call({
+    String? deviceIdentifier = '',
+  }) async {
+    return ApiManager.instance.makeApiCall(
+      callName: 'getContests VTwo',
+      apiUrl: '${ApiGroup.baseUrl}Contest/ByDevice',
+      callType: ApiCallType.GET,
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      params: {
+        'deviceIdentifier': deviceIdentifier,
+      },
+      returnBody: true,
+      encodeBodyUtf8: false,
+      decodeUtf8: false,
+      cache: false,
+      alwaysAllowBody: false,
+    );
+  }
+
+  dynamic amountOfRaces(dynamic response) => getJsonField(
+        response,
+        r'''$.data.count''',
+      );
+  List? contests(dynamic response) => getJsonField(
+        response,
+        r'''$.data.contests''',
+        true,
+      ) as List?;
+  dynamic achievement(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement''',
+      );
+  dynamic achievementLevel(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.level.level''',
+      );
+  dynamic achievementBadge(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.level.badge''',
+      );
+  dynamic achievementLevelName(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.level.name''',
+      );
+  dynamic achievementValue(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.value''',
+      );
+  dynamic achievementMax(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievement.levelCount''',
+      );
 }
 
 class DashboardGraphCall {
@@ -368,19 +473,20 @@ class DashboardGraphCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
-  dynamic years(dynamic response) => getJsonField(
+  List? years(dynamic response) => getJsonField(
         response,
         r'''$.data.years''',
         true,
-      );
-  dynamic points(dynamic response) => getJsonField(
+      ) as List?;
+  List? points(dynamic response) => getJsonField(
         response,
         r'''$.data.points''',
         true,
-      );
+      ) as List?;
 }
 
 class UpdateUserPropertyCall {
@@ -406,6 +512,7 @@ class UpdateUserPropertyCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 }
@@ -439,6 +546,7 @@ class UpdateUserDistancePropertiesCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
@@ -477,6 +585,7 @@ class UpdateUserStrokePropertiesCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
@@ -504,18 +613,19 @@ class GetContestsCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
-  dynamic amount(dynamic response) => getJsonField(
+  int? amount(dynamic response) => castToType<int>(getJsonField(
         response,
         r'''$.data.count''',
-      );
-  dynamic contests(dynamic response) => getJsonField(
+      ));
+  List? contests(dynamic response) => getJsonField(
         response,
         r'''$.data.contests''',
         true,
-      );
+      ) as List?;
 }
 
 class GetContestCall {
@@ -540,18 +650,19 @@ class GetContestCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
-  dynamic clubPerformance(dynamic response) => getJsonField(
+  double? clubPerformance(dynamic response) => castToType<double>(getJsonField(
         response,
         r'''$.data.clubPerformance''',
-      );
-  dynamic clubSwimmers(dynamic response) => getJsonField(
+      ));
+  List? clubSwimmers(dynamic response) => getJsonField(
         response,
         r'''$.data.users''',
         true,
-      );
+      ) as List?;
   dynamic currentUser(dynamic response) => getJsonField(
         response,
         r'''$.data.active''',
@@ -588,13 +699,14 @@ class GetRecomendationsCountCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
-  dynamic amount(dynamic response) => getJsonField(
+  int? amount(dynamic response) => castToType<int>(getJsonField(
         response,
         r'''$.data.count''',
-      );
+      ));
   dynamic unreadAmount(dynamic response) => getJsonField(
         response,
         r'''$.data.unreadCount''',
@@ -624,6 +736,7 @@ class MarkRecomendationAsReadCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 }
@@ -652,6 +765,7 @@ class SwimTimeToClubIdCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
@@ -665,11 +779,108 @@ class SwimTimeToClubIdCall {
       );
 }
 
+class GetAchievementCategoriesCall {
+  Future<ApiCallResponse> call({
+    String? deviceIdentifier = '',
+  }) async {
+    return ApiManager.instance.makeApiCall(
+      callName: 'getAchievementCategories',
+      apiUrl: '${ApiGroup.baseUrl}Achievement/Categories',
+      callType: ApiCallType.GET,
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      params: {
+        'deviceIdentifier': deviceIdentifier,
+      },
+      returnBody: true,
+      encodeBodyUtf8: false,
+      decodeUtf8: false,
+      cache: false,
+      alwaysAllowBody: false,
+    );
+  }
+
+  dynamic myTotalAchievementLevel(dynamic response) => getJsonField(
+        response,
+        r'''$.data.myTotalAchievementLevel''',
+      );
+  dynamic totalAchievementsAvailable(dynamic response) => getJsonField(
+        response,
+        r'''$.data.totalAchievementMaxLevel''',
+      );
+  List? categories(dynamic response) => getJsonField(
+        response,
+        r'''$.data.categories''',
+        true,
+      ) as List?;
+  dynamic categoryCount(dynamic response) => getJsonField(
+        response,
+        r'''$.data.categoryCount''',
+      );
+}
+
+class GetAchievementsInCategoryCall {
+  Future<ApiCallResponse> call({
+    String? deviceIdentifier = '',
+    String? categoryId = '',
+  }) async {
+    return ApiManager.instance.makeApiCall(
+      callName: 'GetAchievementsInCategory',
+      apiUrl: '${ApiGroup.baseUrl}Achievement/Category',
+      callType: ApiCallType.GET,
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      params: {
+        'deviceIdentifier': deviceIdentifier,
+        'categoryId': categoryId,
+      },
+      returnBody: true,
+      encodeBodyUtf8: false,
+      decodeUtf8: false,
+      cache: false,
+      alwaysAllowBody: false,
+    );
+  }
+
+  List? achievements(dynamic response) => getJsonField(
+        response,
+        r'''$.data.achievements''',
+        true,
+      ) as List?;
+}
+
+class PremiumStateCall {
+  Future<ApiCallResponse> call() async {
+    return ApiManager.instance.makeApiCall(
+      callName: 'premiumState',
+      apiUrl: '${ApiGroup.baseUrl}Dashboard/Get/Premium',
+      callType: ApiCallType.GET,
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      params: {},
+      returnBody: true,
+      encodeBodyUtf8: false,
+      decodeUtf8: false,
+      cache: false,
+      alwaysAllowBody: false,
+    );
+  }
+
+  dynamic premiumAllowed(dynamic response) => getJsonField(
+        response,
+        r'''$.data.premium''',
+      );
+}
+
 /// End API Group Code
 
 class GetUserByDeviceIdentifierCall {
   static Future<ApiCallResponse> call({
     String? deviceidentifier = '',
+    bool? resetActiveToMe,
   }) async {
     return ApiManager.instance.makeApiCall(
       callName: 'Get User By DeviceIdentifier',
@@ -678,11 +889,13 @@ class GetUserByDeviceIdentifierCall {
       headers: {},
       params: {
         'deviceIdentifier': deviceidentifier,
+        'resetActiveToMe': resetActiveToMe,
       },
       returnBody: true,
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 }
@@ -703,18 +916,19 @@ class VerifyFullNameCall {
       encodeBodyUtf8: false,
       decodeUtf8: false,
       cache: false,
+      alwaysAllowBody: false,
     );
   }
 
-  static dynamic users(dynamic response) => getJsonField(
+  static List? users(dynamic response) => getJsonField(
         response,
         r'''$.data.users''',
         true,
-      );
-  static dynamic userCount(dynamic response) => getJsonField(
+      ) as List?;
+  static int? userCount(dynamic response) => castToType<int>(getJsonField(
         response,
         r'''$.data.userCount''',
-      );
+      ));
   static dynamic userId(dynamic response) => getJsonField(
         response,
         r'''$.data.users[:].identifier''',
