@@ -149,25 +149,28 @@ class ApiManager {
     BodyType? bodyType,
     bool returnBody,
     bool encodeBodyUtf8,
-    bool decodeUtf8, {
+    bool decodeUtf8,
+    bool alwaysAllowBody, {
     http.Client? client,
   }) async {
     assert(
-      {ApiCallType.POST, ApiCallType.PUT, ApiCallType.PATCH}.contains(type),
+      {ApiCallType.POST, ApiCallType.PUT, ApiCallType.PATCH}.contains(type) ||
+          (alwaysAllowBody && type == ApiCallType.DELETE),
       'Invalid ApiCallType $type for request with body',
     );
     final postBody =
         createBody(headers, params, body, bodyType, encodeBodyUtf8);
 
     if (bodyType == BodyType.MULTIPART) {
-      return multipartRequest(
-          type, apiUrl, headers, params, returnBody, decodeUtf8);
+      return multipartRequest(type, apiUrl, headers, params, returnBody,
+          decodeUtf8, alwaysAllowBody);
     }
 
     final requestFn = {
       ApiCallType.POST: client != null ? client.post : http.post,
       ApiCallType.PUT: client != null ? client.put : http.put,
       ApiCallType.PATCH: client != null ? client.patch : http.patch,
+      ApiCallType.DELETE: client != null ? client.delete : http.delete,
     }[type]!;
     final response = await requestFn(Uri.parse(apiUrl),
         headers: toStringMap(headers), body: postBody);
@@ -181,9 +184,11 @@ class ApiManager {
     Map<String, dynamic> params,
     bool returnBody,
     bool decodeUtf8,
+    bool alwaysAllowBody,
   ) async {
     assert(
-      {ApiCallType.POST, ApiCallType.PUT, ApiCallType.PATCH}.contains(type),
+      {ApiCallType.POST, ApiCallType.PUT, ApiCallType.PATCH}.contains(type) ||
+          (alwaysAllowBody && type == ApiCallType.DELETE),
       'Invalid ApiCallType $type for request with body',
     );
     isFile(e) =>
@@ -286,6 +291,7 @@ class ApiManager {
     bool encodeBodyUtf8 = false,
     bool decodeUtf8 = false,
     bool cache = false,
+    bool alwaysAllowBody = false,
     http.Client? client,
   }) async {
     final callRecord =
@@ -308,7 +314,6 @@ class ApiManager {
     try {
       switch (callType) {
         case ApiCallType.GET:
-        case ApiCallType.DELETE:
           result = await urlRequest(
             callType,
             apiUrl,
@@ -318,6 +323,31 @@ class ApiManager {
             decodeUtf8,
             client: client,
           );
+          break;
+        case ApiCallType.DELETE:
+          result = alwaysAllowBody
+              ? await requestWithBody(
+                  callType,
+                  apiUrl,
+                  headers,
+                  params,
+                  body,
+                  bodyType,
+                  returnBody,
+                  encodeBodyUtf8,
+                  decodeUtf8,
+                  alwaysAllowBody,
+                  client: client,
+                )
+              : await urlRequest(
+                  callType,
+                  apiUrl,
+                  headers,
+                  params,
+                  returnBody,
+                  decodeUtf8,
+                  client: client,
+                );
           break;
         case ApiCallType.POST:
         case ApiCallType.PUT:
@@ -332,6 +362,7 @@ class ApiManager {
             returnBody,
             encodeBodyUtf8,
             decodeUtf8,
+            alwaysAllowBody,
             client: client,
           );
           break;
