@@ -1,6 +1,7 @@
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 import 'base_header_model.dart';
 export 'base_header_model.dart';
@@ -50,7 +51,7 @@ class _BaseHeaderWidgetState extends State<BaseHeaderWidget> {
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             Align(
-              alignment: const AlignmentDirectional(0.00, 0.00),
+              alignment: const AlignmentDirectional(0.0, 0.0),
               child: Container(
                 width: 200.0,
                 height: 50.0,
@@ -68,13 +69,15 @@ class _BaseHeaderWidgetState extends State<BaseHeaderWidget> {
         ),
         if (FFAppState().premium)
           Align(
-            alignment: const AlignmentDirectional(1.00, 0.00),
+            alignment: const AlignmentDirectional(1.0, 0.0),
             child: InkWell(
               splashColor: Colors.transparent,
               focusColor: Colors.transparent,
               hoverColor: Colors.transparent,
               highlightColor: Colors.transparent,
               onTap: () async {
+                HapticFeedback.selectionClick();
+
                 context.pushNamed('advice');
               },
               child: Container(
@@ -84,7 +87,7 @@ class _BaseHeaderWidgetState extends State<BaseHeaderWidget> {
                 child: Stack(
                   children: [
                     Align(
-                      alignment: const AlignmentDirectional(0.00, 0.00),
+                      alignment: const AlignmentDirectional(0.0, 0.0),
                       child: Icon(
                         Icons.circle_notifications,
                         color: FlutterFlowTheme.of(context).text3,
@@ -93,7 +96,7 @@ class _BaseHeaderWidgetState extends State<BaseHeaderWidget> {
                     ),
                     if (FFAppState().recommendationsCount > 0)
                       Align(
-                        alignment: const AlignmentDirectional(1.00, -1.00),
+                        alignment: const AlignmentDirectional(1.0, -1.0),
                         child: Padding(
                           padding: const EdgeInsetsDirectional.fromSTEB(
                               0.0, 2.0, 2.0, 0.0),
@@ -112,7 +115,7 @@ class _BaseHeaderWidgetState extends State<BaseHeaderWidget> {
                               shape: BoxShape.circle,
                             ),
                             child: Align(
-                              alignment: const AlignmentDirectional(0.00, 0.00),
+                              alignment: const AlignmentDirectional(0.0, 0.0),
                               child: Text(
                                 FFAppState().recommendationsCount.toString(),
                                 textAlign: TextAlign.center,
@@ -140,6 +143,7 @@ class _BaseHeaderWidgetState extends State<BaseHeaderWidget> {
           hoverColor: Colors.transparent,
           highlightColor: Colors.transparent,
           onTap: () async {
+            HapticFeedback.selectionClick();
             await widget.drawerClick?.call();
           },
           child: Container(
