@@ -45,9 +45,9 @@ class _ProfileDistanceSettingsWidgetState
     context.watch<FFAppState>();
 
     return Align(
-      alignment: const AlignmentDirectional(0.00, -1.00),
+      alignment: const AlignmentDirectional(0.0, -1.0),
       child: Padding(
-        padding: const EdgeInsetsDirectional.fromSTEB(16.0, 16.0, 16.0, 16.0),
+        padding: const EdgeInsets.all(16.0),
         child: Container(
           width: double.infinity,
           height: double.infinity,
@@ -158,7 +158,7 @@ class _ProfileDistanceSettingsWidgetState
                 ),
                 if (_model.distancePropertyChanged)
                   Align(
-                    alignment: const AlignmentDirectional(0.00, 1.00),
+                    alignment: const AlignmentDirectional(0.0, 1.0),
                     child: wrapWithModel(
                       model: _model.operationButtonModel,
                       updateCallback: () => setState(() {}),
