@@ -1,23 +1,33 @@
 import '/backend/api_requests/api_calls.dart';
 import '/flutter_flow/flutter_flow_util.dart';
+import '/pages/loaders/loader_component/loader_component_widget.dart';
 import 'rankings_widget.dart' show RankingsWidget;
 import 'package:flutter/material.dart';
 
 class RankingsModel extends FlutterFlowModel<RankingsWidget> {
+  ///  Local state fields for this page.
+
+  bool isLoading = false;
+
   ///  State fields for stateful widgets in this page.
 
   final unfocusNode = FocusNode();
   // Stores action output result for [Backend Call - API (Set active user)] action in Container widget.
   ApiCallResponse? apiResult5z2;
+  // Model for LoaderComponent component.
+  late LoaderComponentModel loaderComponentModel;
 
   /// Initialization and disposal methods.
 
   @override
-  void initState(BuildContext context) {}
+  void initState(BuildContext context) {
+    loaderComponentModel = createModel(context, () => LoaderComponentModel());
+  }
 
   @override
   void dispose() {
     unfocusNode.dispose();
+    loaderComponentModel.dispose();
   }
 
   /// Action blocks are added here.
