@@ -2,7 +2,10 @@ import '/backend/api_requests/api_calls.dart';
 import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
+import '/pages/generic/split_times_display/split_times_display_widget.dart';
+import '/pages/loaders/loader_component/loader_component_widget.dart';
 import '/actions/actions.dart' as action_blocks;
+import 'package:aligned_dialog/aligned_dialog.dart';
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -80,11 +83,12 @@ class _RankingsWidgetState extends State<RankingsWidget> {
             borderWidth: 1.0,
             buttonSize: 60.0,
             icon: const Icon(
-              Icons.arrow_back_rounded,
+              Icons.arrow_left,
               color: Colors.white,
               size: 30.0,
             ),
             onPressed: () async {
+              HapticFeedback.selectionClick();
               context.pop();
             },
           ),
@@ -124,147 +128,177 @@ class _RankingsWidgetState extends State<RankingsWidget> {
                   end: const AlignmentDirectional(0, 1.0),
                 ),
               ),
-              child: FutureBuilder<ApiCallResponse>(
-                future: FFAppState().ranking(
-                  uniqueQueryKey:
-                      '${widget.rankingId?.toString()}${widget.fromRank?.toString()}${widget.toRank?.toString()}',
-                  requestFn: () => ApiGroup.rankingsCall.call(
-                    rankingId: widget.rankingId?.toString(),
-                    fromPlace: widget.fromRank?.toString(),
-                    toPlace: widget.toRank?.toString(),
-                  ),
-                ),
-                builder: (context, snapshot) {
-                  // Customize what your widget looks like when it's loading.
-                  if (!snapshot.hasData) {
-                    return Center(
-                      child: SizedBox(
-                        width: 32.0,
-                        height: 32.0,
-                        child: SpinKitWave(
-                          color: FlutterFlowTheme.of(context).text,
-                          size: 32.0,
-                        ),
+              child: Stack(
+                children: [
+                  FutureBuilder<ApiCallResponse>(
+                    future: FFAppState().ranking(
+                      uniqueQueryKey:
+                          '${widget.rankingId?.toString()}${widget.fromRank?.toString()}${widget.toRank?.toString()}',
+                      requestFn: () => ApiGroup.rankingsCall.call(
+                        rankingId: widget.rankingId?.toString(),
+                        fromPlace: widget.fromRank?.toString(),
+                        toPlace: widget.toRank?.toString(),
                       ),
-                    );
-                  }
-                  final containerRankingsResponse = snapshot.data!;
-                  return Container(
-                    width: double.infinity,
-                    height: double.infinity,
-                    decoration: const BoxDecoration(),
-                    child: Column(
-                      mainAxisSize: MainAxisSize.max,
-                      children: [
-                        Align(
-                          alignment: const AlignmentDirectional(0.00, -1.00),
-                          child: Padding(
-                            padding: const EdgeInsetsDirectional.fromSTEB(
-                                0.0, 16.0, 0.0, 16.0),
-                            child: Text(
-                              valueOrDefault<String>(
-                                widget.title,
-                                'Rang',
-                              ),
-                              style: FlutterFlowTheme.of(context)
-                                  .labelMedium
-                                  .override(
-                                    fontFamily: 'Poppins',
-                                    color: FlutterFlowTheme.of(context).text3,
-                                  ),
+                    ),
+                    builder: (context, snapshot) {
+                      // Customize what your widget looks like when it's loading.
+                      if (!snapshot.hasData) {
+                        return Center(
+                          child: SizedBox(
+                            width: 32.0,
+                            height: 32.0,
+                            child: SpinKitWave(
+                              color: FlutterFlowTheme.of(context).text,
+                              size: 32.0,
                             ),
                           ),
-                        ),
-                        Align(
-                          alignment: const AlignmentDirectional(0.00, -1.00),
-                          child: Text(
-                            'Jou rang',
-                            style: FlutterFlowTheme.of(context)
-                                .labelMedium
-                                .override(
-                                  fontFamily: 'Poppins',
-                                  color: FlutterFlowTheme.of(context).text3,
-                                ),
-                          ),
-                        ),
-                        Padding(
-                          padding: const EdgeInsetsDirectional.fromSTEB(
-                              0.0, 4.0, 0.0, 4.0),
-                          child: Container(
-                            width: double.infinity,
-                            height: 50.0,
-                            decoration: const BoxDecoration(),
-                            child: Stack(
-                              children: [
-                                Container(
-                                  width: double.infinity,
-                                  height: double.infinity,
-                                  decoration: const BoxDecoration(
-                                    color: Color(0x40BE9E44),
+                        );
+                      }
+                      final containerRankingsResponse = snapshot.data!;
+                      return Container(
+                        width: double.infinity,
+                        height: double.infinity,
+                        decoration: const BoxDecoration(),
+                        child: Column(
+                          mainAxisSize: MainAxisSize.max,
+                          children: [
+                            Align(
+                              alignment: const AlignmentDirectional(0.0, -1.0),
+                              child: Padding(
+                                padding: const EdgeInsetsDirectional.fromSTEB(
+                                    0.0, 16.0, 0.0, 16.0),
+                                child: Text(
+                                  valueOrDefault<String>(
+                                    widget.title,
+                                    'Rang',
                                   ),
+                                  style: FlutterFlowTheme.of(context)
+                                      .labelMedium
+                                      .override(
+                                        fontFamily: 'Poppins',
+                                        color:
+                                            FlutterFlowTheme.of(context).text3,
+                                      ),
                                 ),
-                                Padding(
-                                  padding: const EdgeInsetsDirectional.fromSTEB(
-                                      8.0, 0.0, 8.0, 0.0),
-                                  child: Row(
-                                    mainAxisSize: MainAxisSize.max,
-                                    mainAxisAlignment:
-                                        MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Padding(
-                                        padding: const EdgeInsetsDirectional.fromSTEB(
-                                            0.0, 0.0, 8.0, 0.0),
-                                        child: Container(
-                                          width: 35.0,
-                                          height: 35.0,
-                                          decoration: BoxDecoration(
-                                            color: FlutterFlowTheme.of(context)
-                                                .buttonYes,
-                                            shape: BoxShape.circle,
-                                          ),
-                                          child: Align(
-                                            alignment: const AlignmentDirectional(
-                                                0.00, 0.00),
-                                            child: AutoSizeText(
-                                              valueOrDefault<String>(
-                                                widget.myRank?.toString(),
-                                                '1',
-                                              ).maybeHandleOverflow(
-                                                  maxChars: 22),
-                                              style: FlutterFlowTheme.of(
-                                                      context)
-                                                  .bodyMedium
-                                                  .override(
-                                                    fontFamily: 'Poppins',
-                                                    color: FlutterFlowTheme.of(
-                                                            context)
-                                                        .text3,
-                                                    fontSize: 16.0,
-                                                    fontWeight: FontWeight.bold,
-                                                  ),
-                                              minFontSize: 12.0,
+                              ),
+                            ),
+                            Align(
+                              alignment: const AlignmentDirectional(0.0, -1.0),
+                              child: Text(
+                                'Jou rang',
+                                style: FlutterFlowTheme.of(context)
+                                    .labelMedium
+                                    .override(
+                                      fontFamily: 'Poppins',
+                                      color: FlutterFlowTheme.of(context).text3,
+                                    ),
+                              ),
+                            ),
+                            Padding(
+                              padding: const EdgeInsetsDirectional.fromSTEB(
+                                  0.0, 4.0, 0.0, 4.0),
+                              child: Container(
+                                width: double.infinity,
+                                height: 50.0,
+                                decoration: const BoxDecoration(),
+                                child: Stack(
+                                  children: [
+                                    Container(
+                                      width: double.infinity,
+                                      height: double.infinity,
+                                      decoration: const BoxDecoration(
+                                        color: Color(0x40BE9E44),
+                                      ),
+                                    ),
+                                    Padding(
+                                      padding: const EdgeInsetsDirectional.fromSTEB(
+                                          8.0, 0.0, 8.0, 0.0),
+                                      child: Row(
+                                        mainAxisSize: MainAxisSize.max,
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.spaceBetween,
+                                        children: [
+                                          Padding(
+                                            padding:
+                                                const EdgeInsetsDirectional.fromSTEB(
+                                                    0.0, 0.0, 8.0, 0.0),
+                                            child: Container(
+                                              width: 35.0,
+                                              height: 35.0,
+                                              decoration: BoxDecoration(
+                                                color:
+                                                    FlutterFlowTheme.of(context)
+                                                        .buttonYes,
+                                                shape: BoxShape.circle,
+                                              ),
+                                              child: Align(
+                                                alignment: const AlignmentDirectional(
+                                                    0.0, 0.0),
+                                                child: AutoSizeText(
+                                                  valueOrDefault<String>(
+                                                    widget.myRank?.toString(),
+                                                    '1',
+                                                  ).maybeHandleOverflow(
+                                                      maxChars: 22),
+                                                  style: FlutterFlowTheme.of(
+                                                          context)
+                                                      .bodyMedium
+                                                      .override(
+                                                        fontFamily: 'Poppins',
+                                                        color:
+                                                            FlutterFlowTheme.of(
+                                                                    context)
+                                                                .text3,
+                                                        fontSize: 16.0,
+                                                        fontWeight:
+                                                            FontWeight.bold,
+                                                      ),
+                                                  minFontSize: 12.0,
+                                                ),
+                                              ),
                                             ),
                                           ),
-                                        ),
-                                      ),
-                                      Align(
-                                        alignment:
-                                            const AlignmentDirectional(-1.00, 0.00),
-                                        child: Container(
-                                          width: 185.0,
-                                          decoration: const BoxDecoration(),
-                                          child: Column(
-                                            mainAxisSize: MainAxisSize.max,
-                                            mainAxisAlignment:
-                                                MainAxisAlignment.center,
-                                            crossAxisAlignment:
-                                                CrossAxisAlignment.start,
-                                            children: [
-                                              Text(
-                                                getJsonField(
-                                                  FFAppState().user,
-                                                  r'''$.active.fullName''',
-                                                ).toString(),
+                                          Align(
+                                            alignment:
+                                                const AlignmentDirectional(-1.0, 0.0),
+                                            child: Container(
+                                              width: 185.0,
+                                              decoration: const BoxDecoration(),
+                                              child: Column(
+                                                mainAxisSize: MainAxisSize.max,
+                                                mainAxisAlignment:
+                                                    MainAxisAlignment.center,
+                                                crossAxisAlignment:
+                                                    CrossAxisAlignment.start,
+                                                children: [
+                                                  Text(
+                                                    getJsonField(
+                                                      FFAppState().user,
+                                                      r'''$.active.fullName''',
+                                                    ).toString(),
+                                                    style: FlutterFlowTheme.of(
+                                                            context)
+                                                        .bodyMedium
+                                                        .override(
+                                                          fontFamily: 'Poppins',
+                                                          color: FlutterFlowTheme
+                                                                  .of(context)
+                                                              .text2,
+                                                        ),
+                                                  ),
+                                                ],
+                                              ),
+                                            ),
+                                          ),
+                                          Container(
+                                            width: 75.0,
+                                            height: 100.0,
+                                            decoration: const BoxDecoration(),
+                                            child: Align(
+                                              alignment: const AlignmentDirectional(
+                                                  0.0, 0.0),
+                                              child: Text(
+                                                widget.myTime!,
                                                 style:
                                                     FlutterFlowTheme.of(context)
                                                         .bodyMedium
@@ -275,236 +309,321 @@ class _RankingsWidgetState extends State<RankingsWidget> {
                                                               .text2,
                                                         ),
                                               ),
-                                            ],
+                                            ),
                                           ),
-                                        ),
-                                      ),
-                                      Container(
-                                        width: 75.0,
-                                        height: 100.0,
-                                        decoration: const BoxDecoration(),
-                                        child: Align(
-                                          alignment:
-                                              const AlignmentDirectional(0.00, 0.00),
-                                          child: Text(
-                                            widget.myTime!,
-                                            style: FlutterFlowTheme.of(context)
-                                                .bodyMedium
-                                                .override(
-                                                  fontFamily: 'Poppins',
-                                                  color: FlutterFlowTheme.of(
-                                                          context)
-                                                      .text2,
-                                                ),
+                                          Container(
+                                            width: 25.0,
+                                            height: 100.0,
+                                            decoration: const BoxDecoration(),
                                           ),
-                                        ),
+                                        ],
                                       ),
-                                      Container(
-                                        width: 25.0,
-                                        height: 100.0,
-                                        decoration: const BoxDecoration(),
-                                      ),
-                                    ],
-                                  ),
+                                    ),
+                                  ],
                                 ),
-                              ],
+                              ),
                             ),
-                          ),
-                        ),
-                        Divider(
-                          thickness: 1.0,
-                          color: FlutterFlowTheme.of(context).accent4,
-                        ),
-                        Expanded(
-                          child: Builder(
-                            builder: (context) {
-                              final ranking = ApiGroup.rankingsCall
-                                      .results(
-                                        containerRankingsResponse.jsonBody,
-                                      )
-                                      ?.toList() ??
-                                  [];
-                              return ListView.builder(
-                                padding: EdgeInsets.zero,
-                                scrollDirection: Axis.vertical,
-                                itemCount: ranking.length,
-                                itemBuilder: (context, rankingIndex) {
-                                  final rankingItem = ranking[rankingIndex];
-                                  return Padding(
-                                    padding: const EdgeInsetsDirectional.fromSTEB(
-                                        0.0, 4.0, 0.0, 4.0),
-                                    child: InkWell(
-                                      splashColor: Colors.transparent,
-                                      focusColor: Colors.transparent,
-                                      hoverColor: Colors.transparent,
-                                      highlightColor: Colors.transparent,
-                                      onTap: () async {
-                                        var shouldSetState = false;
-                                        _model.apiResult5z2 = await ApiGroup
-                                            .setActiveUserCall
-                                            .call(
-                                          deviceIdentifier:
-                                              FFAppState().deviceIdentifier,
-                                          swimrankingsIdentifier: getJsonField(
-                                            rankingItem,
-                                            r'''$.athleteId''',
-                                          ).toString(),
-                                        );
-                                        shouldSetState = true;
-                                        if ((_model.apiResult5z2?.succeeded ??
-                                            true)) {
-                                          await action_blocks
-                                              .getUserAuth(context);
-
-                                          context.pushNamed('Dashboard');
-                                        } else {
-                                          await showDialog(
-                                            context: context,
-                                            builder: (alertDialogContext) {
-                                              return AlertDialog(
-                                                title: const Text('Error'),
-                                                content: const Text(
-                                                    'Kon momenteel de gekozen zwemmer niet inzien! probeer het later opnieuw!'),
-                                                actions: [
-                                                  TextButton(
-                                                    onPressed: () =>
-                                                        Navigator.pop(
-                                                            alertDialogContext),
-                                                    child: const Text('Ok'),
-                                                  ),
-                                                ],
-                                              );
-                                            },
-                                          );
-                                          if (shouldSetState) setState(() {});
-                                          return;
-                                        }
-
-                                        if (shouldSetState) setState(() {});
-                                      },
-                                      child: Container(
-                                        width: 100.0,
-                                        height: 50.0,
-                                        decoration: const BoxDecoration(),
-                                        child: Stack(
-                                          children: [
-                                            if (getJsonField(
+                            Divider(
+                              thickness: 1.0,
+                              color: FlutterFlowTheme.of(context).accent4,
+                            ),
+                            Expanded(
+                              child: Builder(
+                                builder: (context) {
+                                  final ranking = ApiGroup.rankingsCall
+                                          .results(
+                                            containerRankingsResponse.jsonBody,
+                                          )
+                                          ?.toList() ??
+                                      [];
+                                  return ListView.builder(
+                                    padding: EdgeInsets.zero,
+                                    scrollDirection: Axis.vertical,
+                                    itemCount: ranking.length,
+                                    itemBuilder: (context, rankingIndex) {
+                                      final rankingItem = ranking[rankingIndex];
+                                      return Builder(
+                                        builder: (context) => Padding(
+                                          padding:
+                                              const EdgeInsetsDirectional.fromSTEB(
+                                                  0.0, 4.0, 0.0, 4.0),
+                                          child: InkWell(
+                                            splashColor: Colors.transparent,
+                                            focusColor: Colors.transparent,
+                                            hoverColor: Colors.transparent,
+                                            highlightColor: Colors.transparent,
+                                            onTap: () async {
+                                              HapticFeedback.selectionClick();
+                                              setState(() {
+                                                _model.isLoading = true;
+                                              });
+                                              _model.apiResult5z2 =
+                                                  await ApiGroup
+                                                      .setActiveUserCall
+                                                      .call(
+                                                deviceIdentifier: FFAppState()
+                                                    .deviceIdentifier,
+                                                swimrankingsIdentifier:
+                                                    getJsonField(
                                                   rankingItem,
                                                   r'''$.athleteId''',
-                                                ) ==
-                                                getJsonField(
-                                                  FFAppState().user,
-                                                  r'''$.active.athleteId''',
-                                                ))
-                                              Container(
-                                                width: double.infinity,
-                                                height: double.infinity,
-                                                decoration: const BoxDecoration(
-                                                  color: Color(0x40BE9E44),
-                                                ),
-                                              ),
-                                            Padding(
-                                              padding: const EdgeInsetsDirectional
-                                                  .fromSTEB(8.0, 0.0, 8.0, 0.0),
-                                              child: Row(
-                                                mainAxisSize: MainAxisSize.max,
-                                                mainAxisAlignment:
-                                                    MainAxisAlignment
-                                                        .spaceBetween,
+                                                ).toString(),
+                                              );
+                                              if ((_model.apiResult5z2
+                                                      ?.succeeded ??
+                                                  true)) {
+                                                await action_blocks
+                                                    .getUserAuth(context);
+
+                                                context.pushNamed(
+                                                    'personalRecords');
+                                              } else {
+                                                HapticFeedback.heavyImpact();
+                                                await showDialog(
+                                                  context: context,
+                                                  builder:
+                                                      (alertDialogContext) {
+                                                    return AlertDialog(
+                                                      title: const Text('Error'),
+                                                      content: const Text(
+                                                          'Kon momenteel de gekozen zwemmer niet inzien! probeer het later opnieuw!'),
+                                                      actions: [
+                                                        TextButton(
+                                                          onPressed: () =>
+                                                              Navigator.pop(
+                                                                  alertDialogContext),
+                                                          child: const Text('Ok'),
+                                                        ),
+                                                      ],
+                                                    );
+                                                  },
+                                                );
+                                              }
+
+                                              setState(() {
+                                                _model.isLoading = false;
+                                              });
+
+                                              setState(() {});
+                                            },
+                                            onLongPress: () async {
+                                              if (getJsonField(
+                                                    rankingItem,
+                                                    r'''$.splits''',
+                                                  ) !=
+                                                  null) {
+                                                await showAlignedDialog(
+                                                  context: context,
+                                                  isGlobal: true,
+                                                  avoidOverflow: false,
+                                                  targetAnchor:
+                                                      const AlignmentDirectional(
+                                                              0.0, 0.0)
+                                                          .resolve(
+                                                              Directionality.of(
+                                                                  context)),
+                                                  followerAnchor:
+                                                      const AlignmentDirectional(
+                                                              0.0, 0.0)
+                                                          .resolve(
+                                                              Directionality.of(
+                                                                  context)),
+                                                  builder: (dialogContext) {
+                                                    return Material(
+                                                      color: Colors.transparent,
+                                                      child: GestureDetector(
+                                                        onTap: () => _model
+                                                                .unfocusNode
+                                                                .canRequestFocus
+                                                            ? FocusScope.of(
+                                                                    context)
+                                                                .requestFocus(_model
+                                                                    .unfocusNode)
+                                                            : FocusScope.of(
+                                                                    context)
+                                                                .unfocus(),
+                                                        child:
+                                                            SplitTimesDisplayWidget(
+                                                          splits: getJsonField(
+                                                            rankingItem,
+                                                            r'''$.splits''',
+                                                            true,
+                                                          )!,
+                                                          title: widget.title!,
+                                                        ),
+                                                      ),
+                                                    );
+                                                  },
+                                                ).then(
+                                                    (value) => setState(() {}));
+                                              } else {
+                                                HapticFeedback.heavyImpact();
+                                              }
+                                            },
+                                            child: Container(
+                                              width: 100.0,
+                                              height: 50.0,
+                                              decoration: const BoxDecoration(),
+                                              child: Stack(
                                                 children: [
+                                                  if (getJsonField(
+                                                        rankingItem,
+                                                        r'''$.athleteId''',
+                                                      ) ==
+                                                      getJsonField(
+                                                        FFAppState().user,
+                                                        r'''$.active.athleteId''',
+                                                      ))
+                                                    Container(
+                                                      width: double.infinity,
+                                                      height: double.infinity,
+                                                      decoration: const BoxDecoration(
+                                                        color:
+                                                            Color(0x40BE9E44),
+                                                      ),
+                                                    ),
                                                   Padding(
                                                     padding:
                                                         const EdgeInsetsDirectional
-                                                            .fromSTEB(0.0, 0.0,
+                                                            .fromSTEB(8.0, 0.0,
                                                                 8.0, 0.0),
-                                                    child: Container(
-                                                      width: 35.0,
-                                                      height: 35.0,
-                                                      decoration: BoxDecoration(
-                                                        color:
-                                                            FlutterFlowTheme.of(
-                                                                    context)
-                                                                .buttonYes,
-                                                        shape: BoxShape.circle,
-                                                      ),
-                                                      child: Align(
-                                                        alignment:
-                                                            const AlignmentDirectional(
-                                                                0.00, 0.00),
-                                                        child: AutoSizeText(
-                                                          getJsonField(
-                                                            rankingItem,
-                                                            r'''$.place''',
-                                                          )
-                                                              .toString()
-                                                              .maybeHandleOverflow(
-                                                                  maxChars: 22),
-                                                          style: FlutterFlowTheme
-                                                                  .of(context)
-                                                              .bodyMedium
-                                                              .override(
-                                                                fontFamily:
-                                                                    'Poppins',
-                                                                color: FlutterFlowTheme.of(
+                                                    child: Row(
+                                                      mainAxisSize:
+                                                          MainAxisSize.max,
+                                                      mainAxisAlignment:
+                                                          MainAxisAlignment
+                                                              .spaceBetween,
+                                                      children: [
+                                                        Padding(
+                                                          padding:
+                                                              const EdgeInsetsDirectional
+                                                                  .fromSTEB(
+                                                                      0.0,
+                                                                      0.0,
+                                                                      8.0,
+                                                                      0.0),
+                                                          child: Container(
+                                                            width: 35.0,
+                                                            height: 35.0,
+                                                            decoration:
+                                                                BoxDecoration(
+                                                              color: FlutterFlowTheme
+                                                                      .of(context)
+                                                                  .buttonYes,
+                                                              shape: BoxShape
+                                                                  .circle,
+                                                            ),
+                                                            child: Align(
+                                                              alignment:
+                                                                  const AlignmentDirectional(
+                                                                      0.0, 0.0),
+                                                              child:
+                                                                  AutoSizeText(
+                                                                getJsonField(
+                                                                  rankingItem,
+                                                                  r'''$.place''',
+                                                                )
+                                                                    .toString()
+                                                                    .maybeHandleOverflow(
+                                                                        maxChars:
+                                                                            22),
+                                                                style: FlutterFlowTheme.of(
                                                                         context)
-                                                                    .text3,
-                                                                fontSize: 16.0,
-                                                                fontWeight:
-                                                                    FontWeight
-                                                                        .bold,
+                                                                    .bodyMedium
+                                                                    .override(
+                                                                      fontFamily:
+                                                                          'Poppins',
+                                                                      color: FlutterFlowTheme.of(
+                                                                              context)
+                                                                          .text3,
+                                                                      fontSize:
+                                                                          16.0,
+                                                                      fontWeight:
+                                                                          FontWeight
+                                                                              .bold,
+                                                                    ),
+                                                                minFontSize:
+                                                                    12.0,
                                                               ),
-                                                          minFontSize: 12.0,
-                                                        ),
-                                                      ),
-                                                    ),
-                                                  ),
-                                                  Align(
-                                                    alignment:
-                                                        const AlignmentDirectional(
-                                                            -1.00, 0.00),
-                                                    child: Container(
-                                                      width: 185.0,
-                                                      decoration:
-                                                          const BoxDecoration(),
-                                                      child: Column(
-                                                        mainAxisSize:
-                                                            MainAxisSize.max,
-                                                        mainAxisAlignment:
-                                                            MainAxisAlignment
-                                                                .center,
-                                                        crossAxisAlignment:
-                                                            CrossAxisAlignment
-                                                                .start,
-                                                        children: [
-                                                          Text(
-                                                            getJsonField(
-                                                              rankingItem,
-                                                              r'''$.fullName''',
-                                                            ).toString(),
-                                                            style: FlutterFlowTheme
-                                                                    .of(context)
-                                                                .bodyMedium
-                                                                .override(
-                                                                  fontFamily:
-                                                                      'Poppins',
-                                                                  color: FlutterFlowTheme.of(
-                                                                          context)
-                                                                      .text2,
-                                                                ),
+                                                            ),
                                                           ),
-                                                          Align(
+                                                        ),
+                                                        Align(
+                                                          alignment:
+                                                              const AlignmentDirectional(
+                                                                  -1.0, 0.0),
+                                                          child: Container(
+                                                            width: 185.0,
+                                                            decoration:
+                                                                const BoxDecoration(),
+                                                            child: Column(
+                                                              mainAxisSize:
+                                                                  MainAxisSize
+                                                                      .max,
+                                                              mainAxisAlignment:
+                                                                  MainAxisAlignment
+                                                                      .center,
+                                                              crossAxisAlignment:
+                                                                  CrossAxisAlignment
+                                                                      .start,
+                                                              children: [
+                                                                Text(
+                                                                  getJsonField(
+                                                                    rankingItem,
+                                                                    r'''$.fullName''',
+                                                                  ).toString(),
+                                                                  style: FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .bodyMedium
+                                                                      .override(
+                                                                        fontFamily:
+                                                                            'Poppins',
+                                                                        color: FlutterFlowTheme.of(context)
+                                                                            .text2,
+                                                                      ),
+                                                                ),
+                                                                Align(
+                                                                  alignment:
+                                                                      const AlignmentDirectional(
+                                                                          -1.0,
+                                                                          0.0),
+                                                                  child: Text(
+                                                                    getJsonField(
+                                                                      rankingItem,
+                                                                      r'''$.date''',
+                                                                    ).toString(),
+                                                                    style: FlutterFlowTheme.of(
+                                                                            context)
+                                                                        .bodySmall
+                                                                        .override(
+                                                                          fontFamily:
+                                                                              'Poppins',
+                                                                          color:
+                                                                              FlutterFlowTheme.of(context).text2,
+                                                                        ),
+                                                                  ),
+                                                                ),
+                                                              ],
+                                                            ),
+                                                          ),
+                                                        ),
+                                                        Container(
+                                                          width: 75.0,
+                                                          height: 100.0,
+                                                          decoration:
+                                                              const BoxDecoration(),
+                                                          child: Align(
                                                             alignment:
                                                                 const AlignmentDirectional(
-                                                                    -1.00,
-                                                                    0.00),
+                                                                    0.0, 0.0),
                                                             child: Text(
                                                               getJsonField(
                                                                 rankingItem,
-                                                                r'''$.date''',
+                                                                r'''$.time''',
                                                               ).toString(),
                                                               style: FlutterFlowTheme
                                                                       .of(context)
-                                                                  .bodySmall
+                                                                  .bodyMedium
                                                                   .override(
                                                                     fontFamily:
                                                                         'Poppins',
@@ -514,67 +633,51 @@ class _RankingsWidgetState extends State<RankingsWidget> {
                                                                   ),
                                                             ),
                                                           ),
-                                                        ],
-                                                      ),
-                                                    ),
-                                                  ),
-                                                  Container(
-                                                    width: 75.0,
-                                                    height: 100.0,
-                                                    decoration: const BoxDecoration(),
-                                                    child: Align(
-                                                      alignment:
-                                                          const AlignmentDirectional(
-                                                              0.00, 0.00),
-                                                      child: Text(
-                                                        getJsonField(
-                                                          rankingItem,
-                                                          r'''$.time''',
-                                                        ).toString(),
-                                                        style:
-                                                            FlutterFlowTheme.of(
-                                                                    context)
-                                                                .bodyMedium
-                                                                .override(
-                                                                  fontFamily:
-                                                                      'Poppins',
-                                                                  color: FlutterFlowTheme.of(
-                                                                          context)
-                                                                      .text2,
-                                                                ),
-                                                      ),
-                                                    ),
-                                                  ),
-                                                  Container(
-                                                    width: 25.0,
-                                                    height: 100.0,
-                                                    decoration: const BoxDecoration(),
-                                                    child: Icon(
-                                                      Icons.chevron_right,
-                                                      color:
-                                                          FlutterFlowTheme.of(
-                                                                  context)
-                                                              .text2,
-                                                      size: 24.0,
+                                                        ),
+                                                        Container(
+                                                          width: 25.0,
+                                                          height: 100.0,
+                                                          decoration:
+                                                              const BoxDecoration(),
+                                                          child: Icon(
+                                                            Icons.chevron_right,
+                                                            color: FlutterFlowTheme
+                                                                    .of(context)
+                                                                .text2,
+                                                            size: 24.0,
+                                                          ),
+                                                        ),
+                                                      ],
                                                     ),
                                                   ),
                                                 ],
                                               ),
                                             ),
-                                          ],
+                                          ),
                                         ),
-                                      ),
-                                    ),
+                                      );
+                                    },
                                   );
                                 },
-                              );
-                            },
-                          ),
+                              ),
+                            ),
+                          ],
                         ),
-                      ],
+                      );
+                    },
+                  ),
+                  if (_model.isLoading)
+                    Container(
+                      width: double.infinity,
+                      height: double.infinity,
+                      decoration: const BoxDecoration(),
+                      child: wrapWithModel(
+                        model: _model.loaderComponentModel,
+                        updateCallback: () => setState(() {}),
+                        child: const LoaderComponentWidget(),
+                      ),
                     ),
-                  );
-                },
+                ],
               ),
             ),
           ),
